lazily compute resources of DocumentStore handles
Instead creating the DocumentStore (and ZIR if needed) whenever the
document gets modified, it is more efficient to compute them lazily.
This will mean that series of text edits without any requests on them
which often occur while typing will be processed quicker.

I've also tested lazily computing the Ast but it turned out that the Ast
was (almost) always needed which made this unnecessary.